requirements.txt
# Offensive Security Toolkit - Python Dependencies
# Install with: pip install -r requirements.txt

# Core HTTP and Network Libraries
requests>=2.31.0
urllib3>=2.0.0
aiohttp>=3.9.0

# Web Scraping and Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1

# Network Analysis and Packet Manipulation
scapy>=2.5.0
pyshark>=0.6.0

# Protocol Implementation and Network Tools
impacket>=0.11.0
paramiko>=3.4.0
dnspython>=2.4.0

# Cryptography and Encoding
cryptography>=41.0.0
pycryptodome>=3.19.0

# Web Application Testing
selenium>=4.15.0
playwright>=1.40.0

# Data Processing and Analysis
pandas>=2.1.0
numpy>=1.26.0

# Command Line Interface
click>=8.1.0
colorama>=0.4.6
rich>=13.7.0
tabulate>=0.9.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0.1
toml>=0.10.2

# Logging and Monitoring
loguru>=0.7.0

# Report Generation
jinja2>=3.1.2
markdown>=3.5.0
reportlab>=4.0.0

# Database Support
sqlalchemy>=2.0.0
pymongo>=4.6.0

# API and Authentication
pyjwt>=2.8.0
requests-oauthlib>=1.3.0

# Testing and Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0

# Security-Specific Tools
shodan>=1.31.0
censys>=2.2.0
python-whois>=0.8.0

# Performance and Concurrency
asyncio>=3.4.3
concurrent-futures>=3.1.0

# Optional: Machine Learning (for AI-enhanced testing)
# scikit-learn>=1.3.0
# tensorflow>=2.15.0
